pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dt-edit-mcp"
version = "0.1.0"
description = "Darktable headless RAW editor via MCP"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.0.0",
    "lxml>=5.0",
    "pillow>=10.0",
    "numpy>=1.26",
    "scikit-image>=0.22",
    "rawpy>=0.18",
]

[project.scripts]
dt-edit-mcp = "dt_edit_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/dt_edit_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
